feat(cli): add os-checker schema
ref: https://github.com/os-checker/os-checker/issues/69

Usage: os-checker <command> [<args>]

Run a collection of checkers targeting Rust crates, and report bad checking results and statistics.

Options:
  --help            display usage information

Commands:
  setup             Set up all rust-toolchains and checkers without running real
                    checkers.
  run               Run checkers on all repos.
  batch             Merge configs and split it into batches. `os-checker batch
                    --config a.json --config b.json --out-dir batch --size 10`
                    will yield multiple json configs in `batch/`, each
                    containing at most 10 repos.
  schema            Generate a JSON schema file used to validate JSON config.
                    i.e. `{{ "$schema": "./schema.json", /* write config with
                    JSON LSP */ }}`.
